backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
PyPDF2==3.0.1
aiohttp==3.9.3
python-multipart==0.0.9
requests==2.31.0
openai==1.12.0
cachetools==5.3.3
ratelimit==2.2.1
typing-extensions==4.9.0
nltk==3.8.1 